// src/generated/sdk/models/programCreate.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * QMS Backend
 * OpenAPI spec version: 0.1.0
 */
import type { ProgramCreateCustomProperties } from "./programCreateCustomProperties";

export interface ProgramCreate {
  /**
   * @minLength 2
   * @maxLength 200
   */
  name: string;
  description?: string | null;
  website?: string | null;
  portfolio_id: number;
  project_template_id?: number | null;
  workspace_type_id?: number | null;
  artifact_type_id?: number | null;
  is_active?: boolean;
  is_default?: boolean;
  custom_properties?: ProgramCreateCustomProperties;
}